package of the new implementation renamed to proxy
