+++ internal/handler/proxy/middlewares/errorhandler/defaults.go
