+++ internal/handler/proxy/middlewares/errorhandler/mocks/error_handler.go
