+++ internal/handler/proxy/middlewares/errorhandler/options.go
