+++ internal/handler/proxy/middlewares/logger/handler_test.go
