+++ internal/handler/proxy/middlewares/prometheus/defaults.go
