+++ internal/handler/proxy/middlewares/prometheus/options_test.go
